rename a ton of stuff and rework the prelude
