+++ benches/profiling.rs
@@ -2,7 +2,8 @@
 extern crate bencher;
 
 extern crate exr;
-use exr::prelude::*;
+use exr::prelude::common::*;
+use exr::prelude::rgba_image as rgb;
 
 use bencher::Bencher;
 use std::fs;
@@ -12,10 +13,10 @@ fn read_single_image(bench: &mut Bencher) {
     let path = "tests/images/valid/custom/crowskull/crow_uncompressed.exr";
 
     bench.iter(||{
-        rgba::ImageInfo::read_pixels_from_file(
+        rgb::ImageInfo::read_pixels_from_file(
             path, read_options::low(),
-            rgba::pixels::create_flattened_f16,
-            rgba::pixels::flattened_pixel_setter()
+            rgb::pixels::create_flattened_f16,
+            rgb::pixels::flattened_pixel_setter()
         ).unwrap();
     })
 }
@@ -25,10 +26,10 @@ fn read_single_image_from_buffer(bench: &mut Bencher) {
     let file = fs::read("tests/images/valid/custom/crowskull/crow_uncompressed.exr").unwrap();
 
     bench.iter(||{
-        rgba::ImageInfo::read_pixels_from_buffered(
+        rgb::ImageInfo::read_pixels_from_buffered(
             std::io::Cursor::new(&file), read_options::low(),
-            rgba::pixels::create_flattened_f16,
-            rgba::pixels::flattened_pixel_setter()
+            rgb::pixels::create_flattened_f16,
+            rgb::pixels::flattened_pixel_setter()
         ).unwrap();
     })
 }
