+++ examples/1_rgb_exr_to_png.rs
@@ -1,31 +1,31 @@
 
-extern crate image;
+extern crate image as png;
 
 // exr imports
 extern crate exr;
-use exr::prelude::*;
+use exr::prelude::rgba_image as rgb_exr;
 
 fn main() {
 
     // read the image from a file and keep only the png buffer
-    let (_info, png_buffer) = rgba::ImageInfo::read_pixels_from_file(
+    let (_info, png_buffer) = rgb_exr::ImageInfo::read_pixels_from_file(
         "tests/images/valid/openexr/MultiResolution/Kapaa.exr",
-        read_options::high(),
+        rgb_exr::read_options::high(),
 
         // how to create an empty png buffer from exr image meta data (used for loading the exr image)
-        |info: &exr::image::rgba::ImageInfo| {
-            image::ImageBuffer::new(
+        |info: &rgb_exr::ImageInfo| {
+            png::ImageBuffer::new(
                 info.resolution.width() as u32,
                 info.resolution.height() as u32
             )
         },
 
         // set each pixel in the png buffer from the exr file
-        |png_pixels: &mut image::RgbaImage, position: Vec2<usize>, pixel: rgba::Pixel| {
+        |png_pixels: &mut png::RgbaImage, position: rgb_exr::Vec2<usize>, pixel: rgb_exr::Pixel| {
             png_pixels.put_pixel(
                 position.x() as u32, position.y() as u32,
 
-                image::Rgba([
+                png::Rgba([
                     tone_map(pixel.red.to_f32()),
                     tone_map(pixel.green.to_f32()),
                     tone_map(pixel.blue.to_f32()),
