+++ examples/2_rgba_exposure_adjust.rs
@@ -1,8 +1,7 @@
 
 // exr imports
 extern crate exr;
-use exr::prelude::*;
-use exr::image::rgba::{Pixel};
+use exr::prelude::rgba_image::*;
 
 /// Read an RGBA image, increase the exposure, and then write it back.
 /// Uses multi-core compression where appropriate.
@@ -16,12 +15,12 @@ fn main() {
     struct CustomPixels { lines: Vec<Vec<[f16; 4]>> };
 
     // read the image from a file
-    let (mut image_info, mut pixels) = rgba::ImageInfo::read_pixels_from_file(
+    let (mut image_info, mut pixels) = ImageInfo::read_pixels_from_file(
         "tests/images/valid/openexr/MultiResolution/Kapaa.exr",
         read_options::high(),
 
         // create our custom image based on the file info
-        |image: &rgba::ImageInfo| {
+        |image: &ImageInfo| {
             println!("loaded image {:#?}", image);
 
             let default_rgba_pixel = [f16::ZERO, f16::ZERO, f16::ZERO, f16::ONE];
@@ -67,7 +66,7 @@ fn main() {
             "tests/images/out/exposure_adjusted.exr", write_options::high(),
             &|position: Vec2<usize>| -> Pixel {
                 let rgba_f16_array: [f16; 4] = pixels.lines[position.y()][position.x()];
-                rgba::Pixel::from(rgba_f16_array)
+                Pixel::from(rgba_f16_array)
             }
         )
         .unwrap();
