+++ examples/5_extract_pngs.rs
@@ -1,9 +1,7 @@
-extern crate exr;
-extern crate image;
+extern crate image as png;
 
-use exr::prelude::*;
-use exr::image::simple::*;
-use exr::math::Vec2;
+extern crate exr;
+use exr::prelude::simple_image::*;
 use std::cmp::Ordering;
 
 /// For each layer in the exr file,
@@ -26,7 +24,7 @@ pub fn main() {
 
     for (layer_index, layer) in image.layers.iter().enumerate() {
         let layer_name = layer.attributes.name.as_ref()
-            .map_or(String::from("main_layer"), attribute::Text::to_string);
+            .map_or(String::from("main_layer"), Text::to_string);
 
         for channel in &layer.channels {
             match &channel.samples {
@@ -63,7 +61,7 @@ pub fn main() {
 
     /// Save raw float data to a PNG file, doing automatic brightness adjustments per channel
     fn save_f32_image_as_png(data: &[f32], size: Vec2<usize>, name: String) {
-        let mut png_buffer = image::GrayImage::new(size.width() as u32, size.height() as u32);
+        let mut png_buffer = png::GrayImage::new(size.width() as u32, size.height() as u32);
         let mut sorted = Vec::from(data);
         sorted.sort_by(|a, b| a.partial_cmp(b).unwrap_or(Ordering::Less));
 
@@ -85,7 +83,7 @@ pub fn main() {
             let v = (v - min_toned) / (max_toned - min_toned);
 
             // TODO does the `image` crate expect gamma corrected data?
-            *pixel = image::Luma([(v.max(0.0).min(1.0) * 255.0) as u8]);
+            *pixel = png::Luma([(v.max(0.0).min(1.0) * 255.0) as u8]);
         }
 
         png_buffer.save(&name).unwrap();
