+++ src/meta/mod.rs
@@ -2,6 +2,7 @@
 //! Describes all meta data possible in an exr file.
 
 pub mod attribute;
+pub mod attributes;
 
 
 use crate::io::*;
@@ -12,9 +13,9 @@ use crate::error::*;
 use std::fs::File;
 use std::io::{BufReader};
 use crate::math::*;
-use std::collections::{HashSet, HashMap};
+use std::collections::{HashSet};
 use std::convert::TryFrom;
-use smallvec::alloc::fmt::Formatter;
+use crate::meta::attributes::{ImageAttributes, LayerAttributes};
 
 
 /// Contains the complete meta data of an exr image.
@@ -122,160 +123,6 @@ pub struct Header {
     pub own_attributes: LayerAttributes,
 }
 
-
-/// Includes mandatory fields like pixel aspect or display window
-/// which must be the same for all layers.
-/// For more attributes, see struct `LayerAttributes`.
-#[derive(Clone, PartialEq, Debug)]
-pub struct ImageAttributes {
-
-    /// The rectangle anywhere in the global infinite 2D space
-    /// that clips all contents of the file.
-    pub display_window: IntRect,
-
-    /// Aspect ratio of each pixel in this header.
-    pub pixel_aspect: f32,
-
-    /// The chromaticities attribute of the image. See the `Chromaticities` type.
-    pub chromaticities: Option<Chromaticities>,
-
-    /// The time code of the image.
-    pub time_code: Option<TimeCode>,
-
-    /// Optional attributes. Contains custom attributes.
-    /// Does not contain the attributes already present in the `ImageAttributes`.
-    /// Contains only attributes that are standardized to be the same for all headers: chromaticities and time codes.
-    pub custom: HashMap<Text, AttributeValue>,
-}
-
-/// Does not include the attributes required for reading the file contents.
-/// Excludes standard fields that must be the same for all headers.
-/// For more attributes, see struct `ImageAttributes`.
-#[derive(Clone, PartialEq)]
-pub struct LayerAttributes {
-
-    /// The name of this layer.
-    /// Required if this file contains deep data or multiple layers.
-    // As this is an attribute value, it is not restricted in length, may even be empty
-    pub name: Option<Text>,
-
-    /// The bottom left corner of the rectangle that positions this layer
-    /// within the global infinite 2D space of the whole file.
-    /// Equals the position of the data window.
-    pub data_position: Vec2<i32>,
-
-    /// Part of the perspective projection. Default should be `(0, 0)`.
-    // TODO same for all layers?
-    pub screen_window_center: Vec2<f32>, // TODO integrate into `list`
-
-    // TODO same for all layers?
-    /// Part of the perspective projection. Default should be `1`.
-    pub screen_window_width: f32, // TODO integrate into `list`
-
-    /// The white luminance of the colors.
-    /// Defines the luminance in candelas per square meter, Nits, of the RGB value `(1, 1, 1)`.
-    // If the chromaticities and the whiteLuminance of an RGB image are
-    // known, then it is possible to convert the image's pixels from RGB
-    // to CIE XYZ tristimulus values (see function RGBtoXYZ() in header
-    // file ImfChromaticities.h).
-    pub white_luminance: Option<f32>,
-
-    /// The adopted neutral of the colors. Specifies the CIE (x,y) frequency coordinates that should
-    /// be considered neutral during color rendering. Pixels in the image
-    /// whose CIE (x,y) frequency coordinates match the adopted neutral value should
-    /// be mapped to neutral values on the given display.
-    pub adopted_neutral: Option<Vec2<f32>>,
-
-    /// Name of the color transform function that is applied for rendering the image.
-    pub rendering_transform: Option<Text>,
-
-    /// Name of the color transform function that computes the look modification of the image.
-    pub look_modification_transform: Option<Text>,
-
-    /// The horizontal density, in pixels per inch.
-    /// The image's vertical output density can be computed using `x_density * pixel_aspect_ratio`.
-    pub x_density: Option<f32>,
-
-    /// Name of the owner.
-    pub owner: Option<Text>,
-
-    /// Additional textual information.
-    pub comments: Option<Text>,
-
-    /// The date of image creation, in `YYYY:MM:DD hh:mm:ss` format.
-    // TODO parse!
-    pub capture_date: Option<Text>,
-
-    /// Time offset from UTC.
-    pub utc_offset: Option<f32>,
-
-    /// Geographical image location.
-    pub longitude: Option<f32>,
-
-    /// Geographical image location.
-    pub latitude: Option<f32>,
-
-    /// Geographical image location.
-    pub altitude: Option<f32>,
-
-    /// Camera focus in meters.
-    pub focus: Option<f32>,
-
-    /// Exposure time in seconds.
-    pub exposure: Option<f32>,
-
-    /// Camera aperture measured in f-stops. Equals the focal length
-    /// of the lens divided by the diameter of the iris opening.
-    pub aperture: Option<f32>,
-
-    /// Iso-speed of the camera sensor.
-    pub iso_speed: Option<f32>,
-
-    /// If this is an environment map, specifies how to interpret it.
-    pub environment_map: Option<EnvironmentMap>,
-
-    /// Identifies film manufacturer, film type, film roll and frame position within the roll.
-    pub key_code: Option<KeyCode>,
-
-    /// Specifies how texture map images are extrapolated.
-    /// Values can be `black`, `clamp`, `periodic`, or `mirror`.
-    pub wrap_modes: Option<Text>,
-
-    /// Frames per second if this is a frame in a sequence.
-    pub frames_per_second: Option<Rational>,
-
-    /// Specifies the view names for multi-view, for example stereo, image files.
-    pub multi_view: Option<Vec<Text>>,
-
-    /// The matrix that transforms 3D points from the world to the camera coordinate space.
-    /// Left-handed coordinate system, y up, z forward.
-    pub world_to_camera: Option<Matrix4x4>,
-
-    /// The matrix that transforms 3D points from the world to the "Normalized Device Coordinate" space.
-    /// Left-handed coordinate system, y up, z forward.
-    pub world_to_normalized_device: Option<Matrix4x4>,
-
-    /// Specifies whether the pixels in a deep image are sorted and non-overlapping.
-    pub deep_image_state: Option<Rational>,
-
-    /// If the image was cropped, contains the original data window.
-    pub original_data_window: Option<IntRect>,
-
-    /// Level of compression in DWA images.
-    pub dwa_compression_level: Option<f32>,
-
-    /// An 8-bit RGBA image representing the rendered image.
-    pub preview: Option<Preview>,
-
-    /// Name of the view, which is probably either `"right"` or `"left"` for a stereoscopic image.
-    pub view: Option<Text>,
-
-    /// Optional attributes. Contains custom attributes.
-    /// Does not contain the attributes already present in the `Header` or `LayerAttributes` struct.
-    /// Does not contain attributes that are standardized to be the same for all layers: no chromaticities and no time codes.
-    pub custom: HashMap<Text, AttributeValue>,
-}
-
 /// A summary of requirements that must be met to read this exr file.
 /// Used to determine whether this file can be read by a given reader.
 /// It includes the OpenEXR version number. This library aims to support version `2.0`.
@@ -366,38 +213,6 @@ impl Blocks {
 
 
 
-impl LayerAttributes {
-
-    /// Create default layer attributes with a data position of zero.
-    pub fn new(layer_name: Text) -> Self {
-        Self {
-            name: Some(layer_name),
-            .. Self::default()
-        }
-    }
-
-    /// Set the data position of this layer.
-    pub fn with_position(self, data_position: Vec2<i32>) -> Self {
-        Self { data_position, ..self }
-    }
-}
-
-impl ImageAttributes {
-
-    /// Create default image attributes with the specified display window size.
-    /// The display window position is set to zero.
-    pub fn new(display_size: impl Into<Vec2<usize>>) -> Self {
-        Self {
-            display_window: IntRect::from_dimensions(display_size),
-            .. Self::default()
-        }
-    }
-
-    /// Set the data position of this layer.
-    pub fn with_display_window(self, display_window: IntRect) -> Self {
-        Self { display_window, ..self }
-    }
-}
 
 
 /// The first four bytes of each exr file.
@@ -767,7 +582,7 @@ impl Header {
     /// - scan line blocks
     /// - unspecified line order
     /// - no custom attributes
-    pub fn new(name: Text, data_size: impl Into<Vec2<usize>>, channels: SmallVec<[Channel; 5]>) -> Self {
+    pub fn new(name: Text, data_size: impl Into<Vec2<usize>>, channels: SmallVec<[ChannelInfo; 5]>) -> Self {
         let data_size: Vec2<usize> = data_size.into();
         let compression = Compression::Uncompressed;
         let blocks = Blocks::ScanLines;
@@ -1024,7 +839,7 @@ impl Header {
                 }
             }
 
-            for &reserved in attribute::required_attribute_names::ALL.iter() {
+            for &reserved in attributes::standard_names::ALL.iter() {
                 let name  = Text::from_bytes_unchecked(SmallVec::from_slice(reserved));
                 if self.own_attributes.custom.contains_key(&name) || self.shared_attributes.custom.contains_key(&name) {
                     return Err(Error::invalid(format!(
@@ -1114,7 +929,7 @@ impl Header {
             // if the attribute value itself is ok, record it
             match value {
                 Ok(value) => {
-                    use crate::meta::attribute::required_attribute_names as name;
+                    use crate::meta::attributes::standard_names as name;
                     use crate::meta::attribute::AttributeValue::*;
 
                     // if the attribute is a required attribute, set the corresponding variable directly.
@@ -1268,7 +1083,7 @@ impl Header {
         }
 
         {
-            use crate::meta::attribute::required_attribute_names::*;
+            use crate::meta::attributes::standard_names::*;
             use AttributeValue::*;
 
             let (block_type, tiles) = match self.blocks {
@@ -1477,104 +1292,11 @@ impl Requirements {
     }
 }
 
-impl Default for LayerAttributes {
-    fn default() -> Self {
-        Self {
-            data_position: Vec2(0, 0),
-            screen_window_center: Vec2(0.0, 0.0),
-            screen_window_width: 1.0,
-            name: None,
-            white_luminance: None,
-            adopted_neutral: None,
-            rendering_transform: None,
-            look_modification_transform: None,
-            x_density: None,
-            owner: None,
-            comments: None,
-            capture_date: None,
-            utc_offset: None,
-            longitude: None,
-            latitude: None,
-            altitude: None,
-            focus: None,
-            exposure: None,
-            aperture: None,
-            iso_speed: None,
-            environment_map: None,
-            key_code: None,
-            wrap_modes: None,
-            frames_per_second: None,
-            multi_view: None,
-            world_to_camera: None,
-            world_to_normalized_device: None,
-            deep_image_state: None,
-            original_data_window: None,
-            dwa_compression_level: None,
-            preview: None,
-            view: None,
-            custom: Default::default()
-        }
-    }
-}
-
-impl std::fmt::Debug for LayerAttributes {
-    fn fmt(&self, formatter: &mut Formatter<'_>) -> std::fmt::Result {
-        let default_self = Self::default();
-
-        let mut debug = formatter.debug_struct("LayerAttributes (only relevant attributes)");
-
-        // always debug the following fields
-        debug.field("data_position", &self.data_position);
-        debug.field("name", &self.name);
-
-        macro_rules! debug_non_default_fields {
-            ( $( $name: ident ),* ) => { $(
-
-                if self.$name != default_self.$name {
-                    debug.field(stringify!($name), &self.$name);
-                }
-
-            )* };
-        }
-
-        // only debug these fields if they are not the default value
-        debug_non_default_fields! {
-            screen_window_center, screen_window_width,
-            white_luminance, adopted_neutral, x_density,
-            rendering_transform, look_modification_transform,
-            owner, comments,
-            capture_date, utc_offset,
-            longitude, latitude, altitude,
-            focus, exposure, aperture, iso_speed,
-            environment_map, key_code, wrap_modes,
-            frames_per_second, multi_view,
-            world_to_camera, world_to_normalized_device,
-            deep_image_state, original_data_window,
-            dwa_compression_level,
-            preview, view,
-            custom
-        }
-
-        debug.finish()
-    }
-}
-
-impl Default for ImageAttributes {
-    fn default() -> Self {
-        Self {
-            pixel_aspect: 1.0,
-            chromaticities: None,
-            time_code: None,
-            custom: Default::default(),
-            display_window: Default::default(),
-        }
-    }
-}
 
 #[cfg(test)]
 mod test {
     use crate::meta::{MetaData, Requirements, Header, ImageAttributes, LayerAttributes, compute_chunk_count};
-    use crate::meta::attribute::{Text, ChannelList, IntRect, LineOrder, Channel, SampleType};
+    use crate::meta::attribute::{Text, ChannelList, IntRect, LineOrder, ChannelInfo, SampleType};
     use crate::compression::Compression;
     use crate::meta::Blocks;
     use crate::math::*;
@@ -1600,7 +1322,7 @@ mod test {
         let header = Header {
             channels: ChannelList {
                 list: smallvec![
-                    Channel {
+                    ChannelInfo {
                         name: Text::from("main").unwrap(),
                         sample_type: SampleType::U32,
                         quantize_linearly: false,
