+++ tests/dev.rs
@@ -3,13 +3,15 @@
 extern crate exr;
 extern crate smallvec;
 
+use exr::prelude::common::*;
+
 use std::path::{PathBuf};
 use std::ffi::OsStr;
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
-use exr::meta::{MetaData, Header};
+use exr::meta::{Header};
 use std::io;
-use exr::prelude::*;
 use std::io::{Write, Cursor};
+use exr::image::rgba;
 
 fn exr_files() -> impl Iterator<Item=PathBuf> {
     walkdir::WalkDir::new("tests/images/valid").into_iter().map(std::result::Result::unwrap)
